+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/InputMedia/CaptionedInputMedia.kt
@@ -1,9 +1,0 @@
-package com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia
-
-import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput
-
-@Deprecated(
-    "Deprecated due to replace by common type CaptionedOutput",
-    ReplaceWith("CaptionedOutput", "com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.CaptionedOutput")
-)
-typealias CaptionedInputMedia = CaptionedOutput